State auto red up to date with Blue
