+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto/Comp_Autos/Preload/CycleMethods.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Auto.Comp_Autos.Stack_2;
+package org.firstinspires.ftc.teamcode.Auto.Comp_Autos.Preload;
 
 import static java.lang.Thread.sleep;
