+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto/OtherAuto/Old/Stack_2/Blue_Left_Stack_2.java
@@ -1,11 +1,13 @@
-package org.firstinspires.ftc.teamcode.Auto.Comp_Autos.Preload;
+package org.firstinspires.ftc.teamcode.Auto.OtherAuto.Old.Stack_2;
 
 import static org.firstinspires.ftc.teamcode.Constants_and_Setpoints.Constants.propPos;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
+import org.firstinspires.ftc.teamcode.Auto.Comp_Autos.Preload.CycleMethods;
 import org.firstinspires.ftc.teamcode.Odometry.ObjectAvoidance.old.Vector2D;
 import org.firstinspires.ftc.teamcode.Odometry.Pathing.Follower.mecanumFollower;
 import org.firstinspires.ftc.teamcode.Odometry.Pathing.PathGeneration.pathBuilderSubClasses.blueLeftBuilder;
@@ -13,13 +15,15 @@ import org.firstinspires.ftc.teamcode.Odometry.Pathing.PathGeneration.pathBuilde
 import org.firstinspires.ftc.teamcode.Odometry.Pathing.PathGeneration.pathBuilderSubClasses.redRightBuilder;
 import org.firstinspires.ftc.teamcode.VisionTesting.VisionPortalProcessers.propDetectionByAmount;
 import org.firstinspires.ftc.teamcode.hardware._.Collection;
+import org.firstinspires.ftc.teamcode.hardware._.Delivery;
 import org.firstinspires.ftc.teamcode.hardware._.Drivetrain;
 import org.firstinspires.ftc.teamcode.hardware._.Odometry;
 import org.firstinspires.ftc.vision.VisionPortal;
 
-@Autonomous(name = "Blue_Left_Preload", group = "Preload")
+@Autonomous(name = "Blue_Left_Stack+2", group = "Stack 2+2")
+@Disabled
 /**start red right*/
-public class Blue_Left_Preload extends LinearOpMode implements Auto_Methods {
+public class Blue_Left_Stack_2 extends LinearOpMode implements CycleMethods {
 
     public WebcamName frontCam;
 
@@ -37,7 +41,9 @@ public class Blue_Left_Preload extends LinearOpMode implements Auto_Methods {
 
     blueLeftBuilder secondPath = new blueLeftBuilder();
 
-    blueLeftBuilder thirdPath = new blueLeftBuilder();
+    blueLeftBuilder collect = new blueLeftBuilder();
+
+    blueLeftBuilder deliver = new blueLeftBuilder();
 
     blueRightBuilder lastToBackboard = new blueRightBuilder();
 
@@ -50,8 +56,7 @@ public class Blue_Left_Preload extends LinearOpMode implements Auto_Methods {
 
         waitForStart();
 
-        collection.setIntakeHeight(Collection.intakeHeightState.letClawThrough);
-        collection.updateIntakeHeight();
+        autoTimer.reset();
 
         if (propPos == 1){
 
@@ -65,15 +70,33 @@ public class Blue_Left_Preload extends LinearOpMode implements Auto_Methods {
 
             odometry.update();
 
-            Vector2D startPos = new Vector2D(odometry.X, odometry.Y);
+            dropYellowPixel(telemetry);
 
-            dropYellowPixel();
+            collect.buildPath(blueLeftBuilder.Section.collect);
 
-            lastToBackboard.buildPathLine(startPos, new Vector2D(290, 30));
+            deliver.buildPath(blueLeftBuilder.Section.deliver);
 
-            follower.setPath(lastToBackboard.followablePath, lastToBackboard.pathingVelocity);
+            follower.setPath(collect.followablePath, collect.pathingVelocity);
 
-            follower.followPath(180, odometry, drive, "yes");
+            delivery.setGripperState(Delivery.GripperState.open);
+            delivery.updateGrippers();
+
+            collection.setIntakeHeight(Collection.intakeHeightState.thirdPixel);
+            collection.updateIntakeHeight();
+
+            follower.followPath(180, odometry, drive, collection, new Vector2D(125, 180));
+
+            collectPixels();
+
+            follower.setPath(deliver.followablePath, deliver.pathingVelocity);
+
+            follower.followPath(180, odometry, drive, delivery, deliverySlides, collection);
+
+            deployArm();
+
+            dropPixels();
+
+            retractWait();
 
         } else if (propPos == 2) {
 
@@ -89,15 +112,31 @@ public class Blue_Left_Preload extends LinearOpMode implements Auto_Methods {
 
             odometry.update();
 
-            Vector2D startPos = new Vector2D(odometry.X, odometry.Y);
+            dropYellowPixel(telemetry);
 
-            dropYellowPixel();
+            collect.buildPath(blueLeftBuilder.Section.collect);
 
-            lastToBackboard.buildPathLine(startPos, new Vector2D(290, 30));
+            deliver.buildPath(blueLeftBuilder.Section.deliver);
 
-            follower.setPath(lastToBackboard.followablePath, lastToBackboard.pathingVelocity);
+            follower.setPath(collect.followablePath, collect.pathingVelocity);
 
-            follower.followPath(180, odometry, drive, "yes");
+            delivery.setGripperState(Delivery.GripperState.open);
+            delivery.updateGrippers();
+
+            collection.setIntakeHeight(Collection.intakeHeightState.thirdPixel);
+            collection.updateIntakeHeight();
+
+            follower.followPathCollection(180, odometry, drive, collection, delivery, deliverySlides, new Vector2D(125, 180), new Vector2D(121, 153), new Vector2D(86, 139));
+
+            follower.setPath(deliver.followablePath, deliver.pathingVelocity);
+
+            follower.followPathCollection(180, odometry, drive, collection, delivery, deliverySlides, new Vector2D(125, 180), new Vector2D(121, 153), new Vector2D(86, 139), true);
+
+            deployArm();
+
+            dropPixels();
+
+            retractWait();
 
         } else if (propPos == 3) {
 
@@ -119,15 +158,33 @@ public class Blue_Left_Preload extends LinearOpMode implements Auto_Methods {
 
             odometry.update();
 
-            Vector2D startPos = new Vector2D(odometry.X, odometry.Y);
+            dropYellowPixel(telemetry);
 
-            dropYellowPixel();
+            collect.buildPath(blueLeftBuilder.Section.collect);
 
-            lastToBackboard.buildPathLine(startPos, new Vector2D(290, 30));
+            deliver.buildPath(blueLeftBuilder.Section.deliver);
 
-            follower.setPath(lastToBackboard.followablePath, lastToBackboard.pathingVelocity);
+            follower.setPath(collect.followablePath, collect.pathingVelocity);
 
-            follower.followPath(180, odometry, drive, "yes");
+            delivery.setGripperState(Delivery.GripperState.open);
+            delivery.updateGrippers();
+
+            collection.setIntakeHeight(Collection.intakeHeightState.thirdPixel);
+            collection.updateIntakeHeight();
+
+            follower.followPath(180, odometry, drive, collection, new Vector2D(125, 180));
+
+            collectPixels();
+
+            follower.setPath(deliver.followablePath, deliver.pathingVelocity);
+
+            follower.followPath(180, odometry, drive, delivery, deliverySlides, collection);
+
+            deployArm();
+
+            dropPixels();
+
+            retractWait();
 
         }
