+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto/OtherAuto/Test_Autos/TestingPhaseAuto.java
@@ -6,9 +6,7 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.firstinspires.ftc.teamcode.Auto.Comp_Autos.Preload.Auto_Methods;
-import org.firstinspires.ftc.teamcode.Auto.Comp_Autos.Stack_2.CycleMethods;
-import org.firstinspires.ftc.teamcode.Odometry.ObjectAvoidance.old.Vector2D;
+import org.firstinspires.ftc.teamcode.Auto.Comp_Autos.Preload.CycleMethods;
 import org.firstinspires.ftc.teamcode.Odometry.Pathing.Follower.mecanumFollower;
 import org.firstinspires.ftc.teamcode.Odometry.Pathing.PathGeneration.pathBuilderSubClasses.blueLeftBuilder;
 import org.firstinspires.ftc.teamcode.Odometry.Pathing.PathGeneration.pathBuilderSubClasses.blueRightBuilder;
